+++ backendApi/src/modules/permissions/permissions.controller.ts
@@ -1,9 +1,9 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, Req, UseGuards } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Req, UseGuards, HttpException, HttpStatus } from '@nestjs/common';
 import { PermissionsService } from './permissions.service';
-import { CreatePermissionDto } from './dto/create-permission.dto';
-import { UpdatePermissionDto } from './dto/update-permission.dto';
+import { CreatePermissionDto, UpdatePermissionDto } from './dto';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
 import { JwtAuthGuard } from '../auth/guards/auth.guard';
+import { AppError } from 'src/common/constants/error';
 
 @ApiBearerAuth()
 @ApiTags('Permissions')
@@ -13,24 +13,38 @@ export class PermissionsController {
 
   @UseGuards(JwtAuthGuard)
   @Post()
-  create(@Body() createPermissionDto: CreatePermissionDto, @Req() request) {
+  async create(@Body() createPermissionDto: CreatePermissionDto, @Req() request) {
     return this.permissionsService.create(createPermissionDto, request.user.user_id);
   }
 
   @Get('all')
-  findAll() {
+  async findAll() {
     return this.permissionsService.findAll();
   }
 
   @UseGuards(JwtAuthGuard)
   @Patch()
-  update(@Body() updatePermissionDto: UpdatePermissionDto, @Req() request) {
+  async update(@Body() updatePermissionDto: UpdatePermissionDto, @Req() request) {
+    let foundPermission = null;
+    if (updatePermissionDto.permission_id) {
+      foundPermission = await this.permissionsService.findOne(updatePermissionDto.permission_id);
+    }
+
+    if (!foundPermission) {
+      throw new HttpException(AppError.PERMISSION_NOT_FOUND, HttpStatus.NOT_FOUND);
+    }
+
     return this.permissionsService.update(updatePermissionDto, request.user.user_id);
   }
 
   @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string, @Req() request) {
+  async remove(@Param('id') id: string, @Req() request) {
+    const foundPermission = await this.permissionsService.findOne(id);
+    if (!foundPermission) {
+      throw new HttpException(AppError.PERMISSION_NOT_FOUND, HttpStatus.NOT_FOUND);
+    }
+
     return this.permissionsService.remove(id, request.user.user_id);
   }
 }
